requirements.txt
playwright==1.46.0
python-dotenv==1.0.1
